requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# XAI Libraries
captum>=0.6.0
shap>=0.42.0
lime>=0.2.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Image Augmentation
albumentations>=1.3.0
opencv-python>=4.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Logging and Tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Configuration
PyYAML>=6.0
omegaconf>=2.3.0

# Progress Bars
tqdm>=4.65.0

# Jupyter Support
jupyter>=1.0.0
ipykernel>=6.23.0
ipywidgets>=8.0.0

# Utilities
scipy>=1.11.0
requests>=2.31.0
